+++ Bug_Tracker_V1.0/WebApi/Controllers/TicketController.cs
@@ -46,8 +46,13 @@ namespace Bug_Tracker_V1._0.Controllers
         }
 
 
-        public async Task<IActionResult> Index(int id, string? sortOrder)
+        public async Task<IActionResult> Index(int id, string? sortOrder, string? searchString)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
+
             ViewData["DateSortParm"] = String.IsNullOrEmpty(sortOrder) ? "Date_Asc" : "";
 
             var project = await _projectFacade.FindProjectById(id);
@@ -89,6 +94,20 @@ namespace Bug_Tracker_V1._0.Controllers
 
             }
 
+            ViewData["CurrentFilter"] = searchString;
+
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                searchString = searchString.ToLower();
+                sortedTickets = sortedTickets.Where(s => s.AssignedTo.ToLower().Contains(searchString)
+                                       || s.CommentCount.ToString().Contains(searchString)
+                                       || s.Id.ToString().ToLower().Contains(searchString)
+                                       || s.Status.ToString().ToLower().Contains(searchString)
+                                       || s.Title.ToString().ToLower().Contains(searchString)
+                                       || s.Type.ToString().ToLower().Contains(searchString)
+                                       || s.Priority.ToString().Contains(searchString)).ToList();
+
+            }
 
             var vm = new DetailViewModel
             {
@@ -97,7 +116,6 @@ namespace Bug_Tracker_V1._0.Controllers
                 {
                     ProjectUsers = await _projectFacade.GetUserSelectListByProjectId(id),
                     ProjectId = id
-
                 },
                 OpenTicketCount = await _projectFacade.GetProjectTicketCountByStatus(id, Status.Open),
                 ClosedTicketCount = await _projectFacade.GetProjectTicketCountByStatus(id, Status.Closed),
@@ -114,6 +132,10 @@ namespace Bug_Tracker_V1._0.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(DetailViewModel vm)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
             // Check if ticket is empty
             if (string.IsNullOrEmpty(vm.TicketCreateVm.Title) || string.IsNullOrEmpty(vm.TicketCreateVm.Description))
                 return RedirectToAction("Index", "Ticket", new { id = vm.Project.ProjectId });
@@ -141,7 +163,10 @@ namespace Bug_Tracker_V1._0.Controllers
 
         public async Task<IActionResult> Details(int id, string? searchString)
         {
-            ViewData["CurrentFilter"] = searchString;
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
 
             var ticket = await _ticketService.FindOne(id);
             var user = await _projectFacade.GetUser(User);
@@ -151,6 +176,7 @@ namespace Bug_Tracker_V1._0.Controllers
             var ticketHistory = ticket.TicketHistory;
             var project = await _projectFacade.FindProjectById(ticket.ProjectId);
 
+            ViewData["CurrentFilter"] = searchString;
 
             if (!String.IsNullOrEmpty(searchString))
             {
@@ -162,7 +188,7 @@ namespace Bug_Tracker_V1._0.Controllers
             }
 
 
-            if (ticket == null) throw new KeyNotFoundException("Project could not be found in the database with the given ID");
+            if (ticket == null) throw new KeyNotFoundException("Ticket could not be found in the database with the given ID");
 
             var vm = new TicketCreateViewModel
             {
@@ -191,6 +217,11 @@ namespace Bug_Tracker_V1._0.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> PostComment(TicketCreateViewModel tVm)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
+
             var ticket = await _ticketService.FindOne(tVm.Ticket.Id);
             var user = await _projectFacade.GetUser(User);
 
@@ -219,6 +250,10 @@ namespace Bug_Tracker_V1._0.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> UpdateTicket(TicketCreateViewModel vm)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
 
             var ticket = await _ticketService.FindOne(vm.Ticket.Id);
             var oldPriority = ticket.Priority;
@@ -294,6 +329,10 @@ namespace Bug_Tracker_V1._0.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditTicket(TicketCreateViewModel vm)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
             var ticket = await _ticketService.FindOne(vm.Ticket.Id);
 
             ticket.Title = vm.Title;
@@ -310,6 +349,11 @@ namespace Bug_Tracker_V1._0.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete(TicketCreateViewModel vm)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
+
             var ticket = await _ticketService.FindOne(vm.Ticket.Id);
             await _projectFacade.RemoveTicketFKs(ticket.Id);
             await _ticketService.Delete(ticket.Id);
@@ -318,6 +362,11 @@ namespace Bug_Tracker_V1._0.Controllers
 
         public async Task<IActionResult> UserTicketList(string? sortOrder)
         {
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index", "Project");
+            }
+
             var user = await _projectFacade.GetUser(User);
             var sortedTickets = await _projectFacade.GetUserTickets(user);
